feat: add covers
The {{cover}} variable is replaced by the corresponding <img/> tag
in both the layouts and the views.

The cover is set by the `cover` variable in a page's config file.
This variable defaults to `assets/cover.*`, so to add a cover,
one just needs to add a cover file in a page's assets

also fixed wrong escape

Fixes #3